reflection effect added
reflection at interface is based on Snell's law

Fresnel Equation is about energy relationship analyzing polarization state.

about the incident angle, need to change Vector3, (0, sin, cos) and so on

Lines are created by children

The width can be changed in the graph, not a slider or number

The interaction point is fixed at y=0

**The experiment result is shown in the video folder.**

The algorithm draft about phenomenon at interface is updated.

need to update:
The normal vector need to be able to rotate
The refraction ray