Snell's law completed with s- and p- polarization
A new result video is added.

When debugging refraction ray, the incident ray needs to be transfer as parameter.

The y and z from (x,y,z) of refraction ray direction has been changed once, then the result becomes correct.

2 intersection points as s- and p-polarization wave are organized.

The materials are modified, and they are all transparent.

The next step is to map the transparency relationship with Fresnel Equation, which is showing how the power of light is changed after reflection and refracction.