fixing .NET runtime bug
Unity 2021.3.26f1 seems to need to be reinstalled, there is a runtime error happening suddenly:

Error:
  An assembly specified in the application dependencies manifest (Microsoft.NETCore.App.deps.json) was not found:
    package: 'runtime.win-x64.Microsoft.NETCore.App', version: '5.0.3-servicing.21072.12'
    path: 'runtimes/win-x64/native/ucrtbase.dll'

I have reinstalled .NET 5.0.3, but it is not useful and this error did not happen in the last few days. Maybe it is due to the lack of the storage of the disk and something is missing.