Meta sdk sample scenes rendered correctly
Now the UI sets are rendering correctly in Meta SDK sample secens. I feel that maybe it is better to use meta all-in-one sdk rather than XR interaction tool kit, since the platform is more compatible, and is also up-to-date with maturer samples and documents online.

I have tried the sample scene in XR interaction toolkit, which also includes UI canvas, but there are the following 2 issues:

1. Some canvas are purpule and not rendered correctly in VR (This might be corrected after switching platform from Windows to Android)
2. Some canvas or buttons can only be rendered in left eye

Now the Meta all-in-one sdk is running correctly after switching platfrom from Windows to android, with Oculus XR installed(not openXR).